dc2_code/requirement.txt
absl-py==2.1.0
accelerate==0.33.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.41.0
bleach==6.1.0
braceexpand==0.1.7
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.28.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
datasets==2.17.1
debugpy==1.8.1
decorator==5.1.1
decord==0.6.0
deepspeed==0.9.5
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
email_validator==2.2.0
environs==11.0.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.103.2
fastapi-cli==0.0.4
fastjsonschema==2.19.1
feedparser==6.0.10
ffmpy==0.3.2
filelock==3.13.1
fire==0.6.0
flash-attn==2.5.8
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
func-timeout==4.3.5
gitdb==4.0.11
GitPython==3.1.42
google-search-results==2.4.2
gradio==4.15.0
gradio_client==0.8.1
griffe==0.40.1
grpcio==1.64.1
h11==0.14.0
hf_transfer==0.1.6
hjson==3.1.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
humanize==4.10.0
icecream==2.1.3
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
interegular==0.3.3
iopath==0.1.10
ipykernel==6.29.2
ipython==8.22.1
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.17
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lagent==0.2.1
lark==1.1.9
lazy_loader==0.3
linkify-it-py==2.0.3
lit==17.0.6
llava==1.1.3
llvmlite==0.43.0
lm-format-enforcer==0.10.3
lmdeploy==0.4.1
lxml==5.1.0
Markdown==3.6
markdown-it-py==2.2.0
markdown2==2.4.12
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
mmengine==0.10.3
mmengine-lite==0.10.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
notebook==7.1.0
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.36.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
orjson==3.9.14
outlines==0.0.46
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
peft==0.4.0
pexpect==4.9.0
phx-class-registry==4.1.0
pillow==10.2.0
pip==24.0
platformdirs==4.2.0
portalocker==2.8.2
progressbar2==4.3.2
prometheus_client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycocoevalcap==1.2
pycocotools==2.0.7
pycountry==24.6.1
pycparser==2.21
pydantic==1.10.8
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-pptx==0.6.23
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
ray==2.32.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.18.0
ruff==0.2.2
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.13.2
selectivesearch==0.4
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.40.5
setproctitle==1.3.3
setuptools==68.2.2
sgmllib3k==1.0.0
shellingham==1.5.4
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
sty==1.0.6
supervision==0.18.0
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.3.0
tifffile==2024.2.12
tiktoken==0.6.0
timeout-decorator==0.5.0
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
torchvision==0.18.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
transformers-stream-generator==0.0.4
triton==2.3.0
typer==0.12.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.0
tzdata==2024.1
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
validators==0.28.1
visual-genome==1.1.1
vllm==0.5.3.post1+cu118
vllm-flash-attn==2.5.9.post1
vlmeval==0.1.0
wandb==0.16.3
watchfiles==0.22.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==1.13
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.3
wget==3.2
wheel==0.41.2
widgetsnbextension==4.0.10
xformers==0.0.23
XlsxWriter==3.2.0
xtuner==0.1.13
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
